requirements.txt
hidapi>=0.14
Pillow>=10.0
pystray>=0.19
flet>=0.80
